:truck: Split screenshot logic
