+++ src/index.spec.js
@@ -6,127 +6,118 @@ const nodeHtmlToImage = require('./index.js')
 
 describe('node-html-to-image', () => {
   beforeEach(() => {
+    rimraf.sync('./generated')
     fs.mkdirSync('./generated')
   })
 
-  afterEach(() => {
-    rimraf.sync('./generated')
-  })
-
-  it('should generate output file', async () => {
-    await nodeHtmlToImage({
-      output: './generated/image.png',
-      html: '<html></html>'
-    })
-
-    expect(fs.existsSync('./generated/image.png')).toBe(true)
-  })
-
-  it('should return a buffer', async () => {
-    const result = await nodeHtmlToImage({
-      html: '<html></html>'
-    })
-
-    expect(result).toBeInstanceOf(Buffer)
-  })
-
-  it('should throw an error if html is not provided', async () => {
-    let error
-    try {
+  describe('single image', () => {
+    it('should generate output file', async () => {
       await nodeHtmlToImage({
         output: './generated/image.png',
+        html: '<html></html>'
       })
-    } catch (e) {
-      error = e
-    }
-    expect(error.message).toEqual(expect.stringContaining('html'))
-  })
 
-  it('should generate an jpeg image', async () => {
-    await nodeHtmlToImage({
-      output: './generated/image.jpg',
-      html: '<html></html>',
-      type: 'jpeg',
+      expect(fs.existsSync('./generated/image.png')).toBe(true)
     })
 
-    expect(fs.existsSync('./generated/image.jpg')).toBe(true)
-  })
+    it('should return a buffer', async () => {
+      const result = await nodeHtmlToImage({
+        html: '<html></html>'
+      })
 
-  it('should put html in output file', async () => {
-    await nodeHtmlToImage({
-      output: './generated/image.png',
-      html: '<html><body>Hello world!</body></html>'
+      expect(result).toBeInstanceOf(Buffer)
     })
 
-    const text = await getTextFromImage('./generated/image.png')
-    expect(text.trim()).toBe('Hello world!')
-  })
-
-  it('should use handlebars to customize content', async () => {
-    await nodeHtmlToImage({
-      output: './generated/image.png',
-      html: '<html><body>Hello {{name}}!</body></html>',
-      content: { name: 'Yvonnick' }
+    it('should throw an error if html is not provided', async () => {
+      let error
+      try {
+        await nodeHtmlToImage({
+          output: './generated/image.png',
+        })
+      } catch (e) {
+        error = e
+      }
+      expect(error.message).toEqual(expect.stringContaining('html'))
     })
 
-    const text = await getTextFromImage('./generated/image.png')
-    expect(text.trim()).toBe('Hello Yvonnick!')
-  })
-})
+    it('should generate an jpeg image', async () => {
+      await nodeHtmlToImage({
+        output: './generated/image.jpg',
+        html: '<html></html>',
+        type: 'jpeg',
+      })
 
-describe('batch', () => {
-  beforeEach(() => {
-    fs.mkdirSync('./generated')
-  })
-
-  afterEach(() => {
-    rimraf.sync('./generated')
-  })
-
-  it('should create two images', async () => {
-    await nodeHtmlToImage({
-      type: 'png',
-      quality: 300,
-      html: '<html><body>Hello {{name}}!</body></html>',
-      content: [{ name: 'Yvonnick', output: './generated/image1.png' }, { name: 'World', output: './generated/image2.png' }]
+      expect(fs.existsSync('./generated/image.jpg')).toBe(true)
     })
 
-    const text1 = await getTextFromImage('./generated/image1.png')
-    expect(text1.trim()).toBe('Hello Yvonnick!')
+    it('should put html in output file', async () => {
+      await nodeHtmlToImage({
+        output: './generated/image.png',
+        html: '<html><body>Hello world!</body></html>'
+      })
 
-    const text2 = await getTextFromImage('./generated/image2.png')
-    expect(text2.trim()).toBe('Hello World!')
-  })
-
-  it('should return two buffers', async () => {
-    const result = await nodeHtmlToImage({
-      type: 'png',
-      quality: 300,
-      html: '<html><body>Hello {{name}}!</body></html>',
-      content: [{ name: 'Yvonnick' }, { name: 'World' }]
+      const text = await getTextFromImage('./generated/image.png')
+      expect(text.trim()).toBe('Hello world!')
     })
 
-    expect(result[0]).toBeInstanceOf(Buffer)
-    expect(result[1]).toBeInstanceOf(Buffer)
+    it('should use handlebars to customize content', async () => {
+      await nodeHtmlToImage({
+        output: './generated/image.png',
+        html: '<html><body>Hello {{name}}!</body></html>',
+        content: { name: 'Yvonnick' }
+      })
+
+      const text = await getTextFromImage('./generated/image.png')
+      expect(text.trim()).toBe('Hello Yvonnick!')
+    })
   })
 
-  it.skip('should handle mass volume well', async () => {
-    jest.setTimeout(60000 * 60)
-    expect.hasAssertions();
-    const NUMBER_OF_IMAGES = 2000;
-    const content = Array.from(Array(NUMBER_OF_IMAGES), (_, i) => ({
-      name: i,
-      output: `./generated/${i}.jpg`,
-    }));
+  describe('batch', () => {
+    it('should create two images', async () => {
+      await nodeHtmlToImage({
+        type: 'png',
+        quality: 300,
+        html: '<html><body>Hello {{name}}!</body></html>',
+        content: [{ name: 'Yvonnick', output: './generated/image1.png' }, { name: 'World', output: './generated/image2.png' }]
+      })
 
-    await nodeHtmlToImage({
-      type: 'png',
-      quality: 300,
-      html: '<html><body>Hello {{name}}!</body></html>',
-      content,
+      const text1 = await getTextFromImage('./generated/image1.png')
+      expect(text1.trim()).toBe('Hello Yvonnick!')
+
+      const text2 = await getTextFromImage('./generated/image2.png')
+      expect(text2.trim()).toBe('Hello World!')
     })
 
-    expect(fs.readdirSync('./generated')).toHaveLength(NUMBER_OF_IMAGES)
+    it('should return two buffers', async () => {
+      const result = await nodeHtmlToImage({
+        type: 'png',
+        quality: 300,
+        html: '<html><body>Hello {{name}}!</body></html>',
+        content: [{ name: 'Yvonnick' }, { name: 'World' }]
+      })
+
+      expect(result[0]).toBeInstanceOf(Buffer)
+      expect(result[1]).toBeInstanceOf(Buffer)
+    })
+
+    it.skip('should handle mass volume well', async () => {
+      jest.setTimeout(60000 * 60)
+      expect.hasAssertions();
+      const NUMBER_OF_IMAGES = 2000;
+      const content = Array.from(Array(NUMBER_OF_IMAGES), (_, i) => ({
+        name: i,
+        output: `./generated/${i}.jpg`,
+      }));
+
+      await nodeHtmlToImage({
+        type: 'png',
+        quality: 300,
+        html: '<html><body>Hello {{name}}!</body></html>',
+        content,
+      })
+
+      expect(fs.readdirSync('./generated')).toHaveLength(NUMBER_OF_IMAGES)
+    })
   })
 })
